# Conflicts:
#	core/embed/upymod/qstrdefsport.h
#	core/src/trezor/enums/MessageType.py
#	core/src/trezor/messages.py
#	python/setup.py
#	python/src/trezorlib/cli/trezorctl.py
#	python/src/trezorlib/messages.py
#	rust/trezor-client/src/protos/generated/messages.rs

+++ core/embed/rust/src/ui/component/map.rs
@@ -72,7 +72,7 @@ impl<T, F> PageMap<T, F> {
 impl<T, F> Component for PageMap<T, F>
 where
     T: Component,
-    F: Fn(usize) -> usize,
+    F: Fn(u16) -> u16,
 {
     type Msg = T::Msg;
